Add tentative definition of what "present time" means.
For this extension to be useful, we'll want to stay as close as
possible to the time the actual image becomes visible to the
user and abstract away the platform-specific or hardware-specific
finer details. We also want something that's precise enough to
be reliable for the type of applications that require present
timing, unlike VK_KHR_present_wait which offers little guarantee
about accuracy or consistency.